--- :w
@@ -1,28 +1,0 @@
-const fs = require('fs');
-var vorpal = require('vorpal')();
-var _ = require('lodash');
-var Promise = require('bluebird');
-
-//Remove default exit
-const exit = vorpal.find('exit');
-if (exit) {
-  exit.remove();
-}
-
-vorpal
-  .command('exit', 'Quit automax')
-  .alias('quit')
-  .action(function (){
-    process.exit(0);
-  });
-
-vorpal
-  .command('execute', 'Execute prewritten script')
-  .action(function(){
-    const commands = fs.readFileSync('tasks').toString().split('\n');
-    const maps =  commands.map(command => vorpal.execSync(`cinco ${command}`))
-  });
-
-module.exports = vorpal;
-
-
